requirements.txt
telebot
python-dotenv
openai